requirements.txt
# Jupyter and notebook dependencies
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipython>=8.12.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Geospatial data handling
geopandas>=0.14.0

# Network analysis
networkx>=3.1

# Visualization
matplotlib>=3.7.0
folium>=0.15.0
PyYAML
ipyleaflet
nbstripout
scipy>=1.10.0
mapclassify>=2.6.0
scikit-learn>=1.0.0
osmnx>=1.0.0
ipympl>=0.9.0
shapely>=2.0.0

